refactor(web): gather build and deploy settings onto one tabbed page
Deployment source, resource defaults, automatic deployments, image
retention and preview environments all lived on General, while Build &
Deploy owned only the build configuration. That split the pipeline across
two pages and left the page that owns `preset`/`directory` silent about the
deployment source that can rewrite them.

Move them under Build & Deploy as four tabs in pipeline order — Source,
Build, Deploy, Previews — with the active tab in the query string so it is
linkable and survives a reload. General keeps what is genuinely project-wide:
identity, trace sharing, error source context and the danger zone.

Monitoring stays on General rather than moving with the rest: performance
metrics and session recording describe what a running deployment reports,
not how it is built or shipped. Splitting them out of the shared
deployment-config form is safe because `update_project_deployment_config`
applies `if let Some(..)` per field, so each card sends only what it owns and
leaves the rest untouched — verified in both directions against a live server.

That split also fixes a bug: the combined form hardcoded
`sessionRecordingEnabled: false` in its defaults instead of reading the
project, so every "Save Configuration" silently turned session recording off.
Each card now reads its own value.
